remove macOS only phrase
This line almost scared me away before I realized that there was an example at the bottom of the readme.

If you would prefer to remove the bullet completely rather than "minimal support", I can make that change too. I chose to add it because deploying to a device is not supported, and there are open issues around debugging.